+++ examples/blinky_rtc.rs
@@ -15,7 +15,6 @@ use panic_halt as _;
 use stm32f1xx_hal::{pac, prelude::*, rtc::Rtc};
 
 use cortex_m_rt::entry;
-use embedded_hal::digital::v2::OutputPin;
 use nb::block;
 
 #[entry]
@@ -43,10 +42,10 @@ fn main() -> ! {
         rtc.set_alarm(5);
         block!(rtc.wait_alarm()).unwrap();
         if led_on {
-            led.set_low().unwrap();
+            led.set_low();
             led_on = false;
         } else {
-            led.set_high().unwrap();
+            led.set_high();
             led_on = true;
         }
     }