+++ examples/delay.rs
@@ -7,7 +7,6 @@
 use panic_halt as _;
 
 use cortex_m_rt::entry;
-use embedded_hal::digital::v2::OutputPin;
 use stm32f1xx_hal::{delay::Delay, pac, prelude::*};
 
 #[entry]
@@ -34,9 +33,9 @@ fn main() -> ! {
     let mut delay = Delay::new(cp.SYST, clocks);
 
     loop {
-        led.set_high().unwrap();
+        led.set_high();
         delay.delay_ms(1_000_u16);
-        led.set_low().unwrap();
+        led.set_low();
         delay.delay_ms(1_000_u16);
     }
 }